Fixed compilation on 64bit.
git-svn-id: http://svn.aircrack-ng.org/trunk@1208 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/airdecloak-ng.c
@@ -400,7 +400,10 @@ BOOLEAN read_packets(void)
 
         if( bytes_read != (size_t) _packet_elt_head->current->header.caplen )
         {
-			printf("Error reading the file: read %d bytes out of %d.\n", bytes_read, _packet_elt_head->current->header.caplen);
+			printf("Error reading the file: read %lu bytes out of %d.\n",
+						(unsigned long) bytes_read,
+						_packet_elt_head->current->header.caplen);
+
             break;
 		}
